chore: Adjusted pages and extracted footer
