+++ src/app/[locale]/how-it-works/page.tsx
