--- src/app/log-in/layout.tsx
@@ -1,10 +1,0 @@
-import type { Metadata } from "next";
-import RootLayout from "../rootLayout";
-
-export const metadata: Metadata = {
-  title: "Log in - Wishlist for friends",
-  description:
-    "This application is a tool that allows users to create and manage lists of desired items or services. These applications are popular among shoppers, gift enthusiasts, and people who want to plan their purchases.",
-};
-
-export default RootLayout;
